+++ tests/testthat/test-class_fst.R
@@ -18,6 +18,15 @@ tar_test("fst format", {
   expect_silent(target_validate(x))
 })
 
+tar_test("non-data-frame throws an error", {
+  skip_if_not_installed("fst")
+  tar_script(
+    tar_target(x, targets::tar_target(x, 1), format = "fst")
+  )
+  expect_error(tar_make(callr_function = NULL), class = "tar_condition_run")
+  expect_gt(nchar(tar_meta(x, error)$error[1]), 5)
+})
+
 tar_test("bad compression level throws error (structured resources)", {
   skip_if_not_installed("fst")
   tar_script({
@@ -62,29 +71,33 @@ tar_test("bad compression level throws error (unstructured resources)", {
   suppressWarnings(
     expect_error(
       tar_make(callr_function = NULL),
-      class = "tar_condition_validate"
+      class = "tar_condition_run"
     )
   )
 })
 
 tar_test("fst packages", {
+  skip_if_not_installed("fst")
   x <- tar_target(x, 1, format = "fst")
   out <- store_get_packages(x$store)
   expect_equal(out, "fst")
 })
 
 tar_test("does not inherit from tar_external", {
+  skip_if_not_installed("fst")
   store <- tar_target(x, "x_value", format = "fst")$store
   expect_false(inherits(store, "tar_external"))
 })
 
 tar_test("store_row_path()", {
+  skip_if_not_installed("fst")
   store <- tar_target(x, "x_value", format = "fst")$store
   store$file$path <- "path"
   expect_equal(store_row_path(store), NA_character_)
 })
 
 tar_test("store_path_from_record()", {
+  skip_if_not_installed("fst")
   store <- tar_target(x, "x_value", format = "fst")$store
   record <- record_init(name = "x", path = "path", format = "fst")
   expect_equal(